Add a controller as the first object to the gamepad callbacks.
This is so that downstream consumers can, at their option,
get all of the data associated with a gamepad when a button
changes.  This is especially useful when trying to generate
e.g. ROS 2 sensor_msgs/msg/Joy messages, which expect the
complete state of all buttons and axes to be published on
every change.

Signed-off-by: Chris Lalancette <clalancette@openrobotics.org>