Don't treat gamepad axes as pairs. (#9)
It turns out that gamepad axes are not truly paired, at least,
not along even/odd pairs.  That is, on my gamepad here, the
left analog left/right is axis 0, and the left analog front/back
is axis 1.  But my right analog left/right is axis 3, and the
right analog front/back is axis 4.  So we shouldn't treat these
as pairs, instead just treat them all as individual axes.

Signed-off-by: Chris Lalancette <clalancette@openrobotics.org>