Add in third-person follow view. (#20)
* Add in third-person follow view.

This makes a view that hovers above the target object, but looks
off into the distance.  It sort of mimics what a drone might do,
but more closesly mimics what video games do.

One thing it does not currently deal with is obstructions in the
scene.  If the camera happens to be behind something in the
scene, then the camera will just be obstructed.  Fixing this is
future work.

Signed-off-by: Chris Lalancette <clalancette@openrobotics.org>