Fix a problem with multiple audio messages coming in at once. (#41)
The problem is that currently, we fetch the asset asynchronously,
and only then add it to the map.  That means that if other
audio messages come in while we are fetching, we'll fetch the
data more than once and then play it more than once.

Instead, this PR changes things so that as soon as we get a
request for a new audio asset, we immediately add it to the
map, then go fetch it asynchronously.  If other audio messages
for the same asset come in, they will just harmlessly try to
play audio that isn't there yet.  When we have finished downloading,
we replace the object in the map with the fully active one
which can be played.

Signed-off-by: Chris Lalancette <clalancette@openrobotics.org>